清理本地archive的工具函数
